# Conflicts:
#	app/src/main/java/com/adrosonic/craftexchange/ui/modules/buyer/enquiry/CompletedEnquiryFragment.kt
#	app/src/main/java/com/adrosonic/craftexchange/ui/modules/buyer/search/BuyerSearchResultFragment.kt

+++ app/src/main/java/com/adrosonic/craftexchange/repository/remote/WishlistDao.kt
@@ -14,7 +14,7 @@ interface WishlistDao {
     @Headers("Accept: application/json")
     @POST("product/addToWishlist/{productId}")
     fun addToWishlist(@Header("Authorization") token:String,
-                      @Path("productId") productId : Int
+                      @Path("productId") productId : Long
     ) : Call<NotificationReadResponse>
 
     @Headers("Accept: application/json")
